refactor: #19 general refactoring
