+++ src/main/java/com/janluk/schoolmanagementapp/common/email/EmailProperties.java
@@ -1,6 +1,7 @@
-package com.janluk.schoolmanagementapp.common.email.config;
+package com.janluk.schoolmanagementapp.common.email;
 
-public class EmailProperties {
+class EmailProperties {
+
     public static final String ACCOUNT_CONFIRM_URI = "/account/confirm";
     public static final String CONFIRMATION_TITLE = "Confirmation of password change";
     public static final String FAKE_SENDER = "fakesender@gmail.com";
