+++ src/main/java/com/janluk/schoolmanagementapp/common/model/RoleEntity.java
@@ -1,19 +1,16 @@
 package com.janluk.schoolmanagementapp.common.model;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import java.io.Serializable;
 
 @Entity
 @Table(name = "roles")
 @Getter
-@Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@EqualsAndHashCode(of = "role")
 public class RoleEntity implements Serializable {
 
     @Id
