+++ src/main/java/com/janluk/schoolmanagementapp/common/model/vo/ClassType.java
@@ -1,8 +1,5 @@
 package com.janluk.schoolmanagementapp.common.model.vo;
 
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
 public enum ClassType {
 
     A1, A2, A3,
