+++ src/main/java/com/janluk/schoolmanagementapp/common/model/vo/SubjectType.java
@@ -1,8 +1,5 @@
 package com.janluk.schoolmanagementapp.common.model.vo;
 
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
 public enum SubjectType {
 
     MATHEMATICS, BIOLOGY, CHEMISTRY, PHYSICS, GEOGRAPHY, ENGLISH
