+++ src/main/java/com/janluk/schoolmanagementapp/common/repository/adapter/SqlUserRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public class SqlUserRepository implements UserRepository {
+class SqlUserRepository implements UserRepository {
 
     private final JpaUserRepository jpaUserRepository;
