+++ src/main/java/com/janluk/schoolmanagementapp/security/config/SecurityConfig.java
@@ -23,13 +23,18 @@ import org.springframework.security.web.SecurityFilterChain;
 @Configuration
 @EnableWebSecurity
 @EnableMethodSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
+class SecurityConfig {
 
     private final CustomizedUserDetailsService userDetailsService;
+    private final String API_PREFIX;
 
-    @Value("${api.prefix}")
-    String API_PREFIX;
+    public SecurityConfig(
+            CustomizedUserDetailsService userDetailsService,
+            @Value("${api.prefix}") String API_PREFIX
+    ) {
+        this.userDetailsService = userDetailsService;
+        this.API_PREFIX = API_PREFIX;
+    }
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
