+++ src/main/java/com/janluk/schoolmanagementapp/security/service/AccountService.java
@@ -1,6 +1,6 @@
 package com.janluk.schoolmanagementapp.security.service;
 
-import com.janluk.schoolmanagementapp.common.email.service.EmailService;
+import com.janluk.schoolmanagementapp.common.email.EmailService;
 import com.janluk.schoolmanagementapp.common.model.UserEntity;
 import com.janluk.schoolmanagementapp.common.repository.port.UserRepository;
 import com.janluk.schoolmanagementapp.security.schema.ConfirmPasswordRequest;
@@ -8,8 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import static com.janluk.schoolmanagementapp.common.user.TokenGenerator.*;
 
 @Service
 @RequiredArgsConstructor
@@ -23,15 +21,14 @@ public class AccountService {
     public void confirmPassword(String token, ConfirmPasswordRequest request) {
         UserEntity user = userRepository.getByPasswordConfirmationToken(token);
 
-        user.setPassword(passwordEncoder.encode(request.password()));
-        user.setPasswordConfirmationToken(null);
+        user.updatePasswordAndClearToken(passwordEncoder.encode(request.password()));
     }
 
     @Transactional
     public void changePassword(String email) {
         UserEntity user = userRepository.getByEmail(email);
 
-        user.setPasswordConfirmationToken(generateToken());
+        user.generateNewToken();
         emailService.sendNotification(email, user.getPasswordConfirmationToken());
 
         userRepository.save(user);
