+++ src/main/java/com/janluk/schoolmanagementapp/student/criteria/StudentFields.java
@@ -1,6 +1,6 @@
 package com.janluk.schoolmanagementapp.student.criteria;
 
-public class StudentFields {
+class StudentFields {
 
     static final String USER = "user";
 }
