+++ src/main/java/com/janluk/schoolmanagementapp/teacher/service/AdminTeacherService.java
@@ -2,23 +2,12 @@ package com.janluk.schoolmanagementapp.teacher.service;
 
 import com.janluk.schoolmanagementapp.common.criteria.CommonUserFilters;
 import com.janluk.schoolmanagementapp.common.exception.EmailAlreadyExistsException;
-import com.janluk.schoolmanagementapp.common.model.SchoolClassEntity;
-import com.janluk.schoolmanagementapp.common.model.SchoolSubjectEntity;
 import com.janluk.schoolmanagementapp.common.model.TeacherEntity;
-import com.janluk.schoolmanagementapp.common.model.vo.SubjectType;
-import com.janluk.schoolmanagementapp.common.repository.port.SchoolClassRepository;
-import com.janluk.schoolmanagementapp.common.repository.port.SchoolSubjectRepository;
 import com.janluk.schoolmanagementapp.common.repository.port.TeacherRepository;
 import com.janluk.schoolmanagementapp.common.schema.CourseDTO;
-import com.janluk.schoolmanagementapp.common.schema.SchoolClassRequest;
-import com.janluk.schoolmanagementapp.common.schema.SchoolSubjectRequest;
 import com.janluk.schoolmanagementapp.common.user.RoleAdder;
 import com.janluk.schoolmanagementapp.common.user.UserValidator;
 import com.janluk.schoolmanagementapp.teacher.criteria.TeacherSearcher;
-import com.janluk.schoolmanagementapp.teacher.exception.TeacherAlreadyTeachingSubjectException;
-import com.janluk.schoolmanagementapp.teacher.exception.TeacherIsAlreadyTutor;
-import com.janluk.schoolmanagementapp.teacher.exception.TeacherNotAssignedAsTutorException;
-import com.janluk.schoolmanagementapp.common.exception.TeacherNotTeachingSubjectException;
 import com.janluk.schoolmanagementapp.teacher.mapper.TeacherMapper;
 import com.janluk.schoolmanagementapp.teacher.schema.CreateTeacherRequest;
 import com.janluk.schoolmanagementapp.teacher.schema.TeacherDTO;
@@ -30,7 +19,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Set;
 import java.util.UUID;
 
 @Service
@@ -40,8 +28,6 @@ public class AdminTeacherService {
 
     private final UserValidator userValidator;
     private final TeacherRepository teacherRepository;
-    private final SchoolClassRepository schoolClassRepository;
-    private final SchoolSubjectRepository schoolSubjectRepository;
     private final TeacherMapper teacherMapper;
     private final RoleAdder roleAdder;
     private final TeacherSearcher teacherSearcher;
@@ -71,84 +57,5 @@ public class AdminTeacherService {
         roleAdder.addTeacherRole(teacher);
 
         return teacherRepository.save(teacher).toString();
-    }
-
-    @Transactional
-    public String assignTutorToTeacher(UUID id, SchoolClassRequest request) {
-        TeacherEntity teacher = teacherRepository.getById(id);
-        SchoolClassEntity schoolClass = schoolClassRepository.getById(request.classType());
-
-        if (isTeacherTutorOfSchoolClass(teacher, schoolClass)) {
-            log.warn(
-                    "Teacher with id %s is already the tutor of class %s."
-                            .formatted(teacher.getId(), request.classType().name())
-            );
-            throw new TeacherIsAlreadyTutor(teacher.getId().toString(), request.classType().name());
-        }
-        teacher.assignTutor(schoolClass);
-
-        return teacher.getId().toString();
-    }
-
-    @Transactional
-    public void removeTutorAssignment(UUID id) {
-        TeacherEntity teacher = teacherRepository.getById(id);
-
-        if (!isTeacherAlreadyTutor(teacher)) {
-            log.warn(
-                    "Teacher with id %s is not assigned as a tutor to any school class."
-                            .formatted(teacher.getId().toString())
-            );
-            throw new TeacherNotAssignedAsTutorException(teacher.getId().toString());
-        }
-
-        teacher.setTutorClass(null);
-    }
-
-    @Transactional
-    public String assignSubjectToTutor(UUID id, SchoolSubjectRequest request) {
-        TeacherEntity teacher = teacherRepository.getById(id);
-        SchoolSubjectEntity schoolSubject = schoolSubjectRepository.getById(request.subjectType());
-
-        if (isTeacherOfSchoolSubject(teacher, schoolSubject)) {
-            log.warn(
-                    "Teacher with id: %s is already teaching subject: %s."
-                            .formatted(teacher.getId().toString(), request.subjectType().name())
-            );
-            throw new TeacherAlreadyTeachingSubjectException(teacher.getId().toString(), request.subjectType().name());
-        }
-
-        teacher.getTaughtSubjects().add(schoolSubject);
-
-        return teacher.getId().toString();
-    }
-
-    @Transactional
-    public void removeSubjectFromTeacher(UUID id, SubjectType subject) {
-        TeacherEntity teacher = teacherRepository.getById(id);
-        SchoolSubjectEntity schoolSubject = schoolSubjectRepository.getById(subject);
-
-        if (!isTeacherOfSchoolSubject(teacher, schoolSubject)) {
-            log.warn("Teacher with id: %s does not teach subject: %s.".formatted(teacher.getId(), subject.name()));
-            throw new TeacherNotTeachingSubjectException(teacher.getId().toString(), subject.name());
-        }
-
-        teacher.getTaughtSubjects().remove(schoolSubject);
-    }
-
-    private boolean isTeacherTutorOfSchoolClass(TeacherEntity teacher, SchoolClassEntity schoolClass) {
-        SchoolClassEntity tutorClass = teacher.getTutorClass();
-
-        return tutorClass != null && tutorClass.equals(schoolClass);
-    }
-
-    private boolean isTeacherAlreadyTutor(TeacherEntity teacher) {
-        return teacher.getTutorClass() != null;
-    }
-
-    private boolean isTeacherOfSchoolSubject(TeacherEntity teacher, SchoolSubjectEntity schoolSubject) {
-        Set<SchoolSubjectEntity> taughtSubjects = teacher.getTaughtSubjects();
-
-        return taughtSubjects.contains(schoolSubject);
     }
 }
