+++ src/main/resources/application-dev.yml
@@ -29,5 +29,5 @@ rsa:
 application:
   url: http://localhost:8080/api/v1
 
-# defines whether to use real email sender or a fake one
-use-email-sender: false
+# defines whether to use real email sender or a fake one. Options: real / fake.
+use-email-sender: fake
